+++ CHANGELOG.md
@@ -24,3 +24,5 @@ All notable changes to this project will be documented in this file. See [standa
 * [#42](https://github.com/upfrontsystems/izinto-angular/issues/42) fix chart date alignment
 * [#49](https://github.com/upfrontsystems/izinto-angular/issues/42) fix data truncation
 * [#62](https://github.com/upfrontsystems/izinto-angular/issues/42) remove template components
+* [#12](https://github.com/upfrontsystems/izinto-angular/issues/12) show wind speed on wind arrow chart
+